+++ lib/turbo_mvvm.dart
@@ -8,7 +8,7 @@ export 'data/mixins/error_management.dart';
 export 'data/mixins/view_model_helpers.dart';
 
 export 'data/models/ake_base_view_model.dart';
-export 'data/models/base_view_model.dart';
+export 'data/models/turbo_view_model.dart';
 export 'data/models/busy_model.dart';
 
 export 'services/busy_service.dart';