+++ test/models/base_view_model_implementation.dart
@@ -1,8 +1,8 @@
-import 'package:turbo_mvvm/data/models/base_view_model.dart';
+import 'package:turbo_mvvm/data/models/turbo_view_model.dart';
 import 'package:turbo_mvvm/data/mixins/busy_management.dart';
 import 'package:turbo_mvvm/data/mixins/error_management.dart';
 
-class BaseViewModelImplementation<T> extends BaseViewModel<T>
+class BaseViewModelImplementation<T> extends TurboViewModel<T>
     with ErrorManagement, BusyManagement {
   BaseViewModelImplementation({
     required bool isMock,